🖖 make sure unit tests aren't doing network calls
Running the tests with the new verbose logging revealed that one
network-calling function hadn't been patched. Added a pytest extension
to fail the run if there's a network call and then added the remaining
required mock.

Fun fact: there is (of course) a GitHub user called "starfleet", and
the test suite was trying to retrieve information about it via the
GitHub API.